% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci2.R
\name{ci2estimate}
\alias{ci2estimate}
\title{Transforms confidence intervals into parameter estimates.}
\usage{
ci2estimate(lower, upper, ratio = FALSE, antilog = FALSE)
}
\arguments{
\item{lower}{Numeric vector of lower confidence interval bounds.}

\item{upper}{Numeric vector of upper confidence interval bounds.}

\item{ratio}{Indicates whether the confidence interval is for a ratio, e.g. an
odds ratio, relative risk or hazard ratio. Defaults to \code{FALSE}.}

\item{antilog}{Indicates whether the estimate is reported on the ratio scale.
Only applies if \code{ratio = TRUE}. Defaults to \code{FALSE}.}
}
\value{
A numeric vector of parameter estimates.
}
\description{
Transforms confidence intervals into parameter estimates.
}
\examples{
ci2estimate(lower = 1, upper = 3)
ci2estimate(lower = 1, upper = 3, ratio = TRUE)
ci2estimate(lower = 1, upper = 3, ratio = TRUE, antilog = TRUE)
}
\seealso{
\code{\link{ci2p}}, \code{\link{ci2z}}, \code{\link{ci2se}}, \code{\link{p2z}}, \code{\link{z2p}}
}
