% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci2.R
\name{ci2p}
\alias{ci2p}
\title{Transforms confidence intervals into p-values.}
\usage{
ci2p(lower, upper, conf.level = 0.95, ratio = FALSE, alternative = "two.sided")
}
\arguments{
\item{lower}{Numeric vector of lower confidence interval bounds.}

\item{upper}{Numeric vector of upper confidence interval bounds.}

\item{conf.level}{The confidence level of the confidence intervals. Default is 0.95.}

\item{ratio}{Indicates whether the confidence interval is for a ratio, e.g. an
odds ratio, relative risk or hazard ratio. Defaults to \code{FALSE}.}

\item{alternative}{Specifies whether the p-values are "two.sided" (default) or "one.sided".}
}
\value{
A numeric vector of p-values.
}
\description{
Transforms confidence intervals into p-values.
}
\examples{
ci2p(lower = 1, upper = 3)
ci2p(lower = 1, upper = 3, alternative = "one.sided")
}
\seealso{
\code{\link{ci2p}}, \code{\link{ci2estimate}}, \code{\link{ci2se}}, \code{\link{p2z}}, \code{\link{z2p}}
}
