% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levelEquivalent.R
\name{levelEquivalent}
\alias{levelEquivalent}
\title{Computes the equivalent threshold for the sceptical p-value}
\usage{
levelEquivalent(dinf, level = 0.025, alternative = c("one.sided", "two.sided"))
}
\arguments{
\item{dinf}{The limiting relative effect size bound below which replication success is
deemed to be impossible for borderline significant original studies
with p-values equal to \code{level}.}

\item{level}{Threshold for the calibrated sceptical p-value.
Default is 0.025.}

\item{alternative}{Specifies if \code{level} is "one.sided" (default) or
 "two.sided". If "one-sided",
then a one-sided equivalent threshold for the sceptical p-value is computed.}
}
\value{
Equivalent thresholds for the sceptical p-value corresponding to \code{dmin}=1.
Thus, at the equivalent threshold, replication success is impossible for borderline significant original
studies and shrinkage of the replication effect estimate.
}
\description{
The equivalent threshold for the sceptical p-value is computed based on
the specified limiting relative effect size, the level, and
the alternative.
}
\examples{
levelEquivalent(dinf = 0.8, level = 0.025)
levelEquivalent(dinf = 0.8, level = 0.05, alternative="two.sided")
}
\references{
Held, L. (2020). A new standard for the analysis and design of replication studies (with discussion).
\emph{Journal of the Royal Statistical Society: Series A (Statistics in Society)}, \bold{183}, 431-448.
\doi{10.1111/rssa.12493}

Held, L., Micheloud, C., Pawel, S. (2022). The assessment of
    replication success based on relative effect size. The Annals of Applied
    Statistics. 16:706-720. \doi{10.1214/21-AOAS1502}
}
\seealso{
\code{\link{pSceptical}}, \code{\link{levelSceptical}}
}
\author{
Leonhard Held
}
