% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversionHelpers.R
\name{p2z}
\alias{p2z}
\title{Transforms p-values to z-values.}
\usage{
p2z(p, alternative = "two.sided")
}
\arguments{
\item{p}{Numeric vector of p-values.}

\item{alternative}{Direction of the alternative of the p-value. 
Either "one.sided" (default), "two.sided", "less", or "greater".
If "one.sided" or "two.sided", the z-value is assumed to be positive.}
}
\value{
A numeric vector of z-values.
}
\description{
Transforms p-values to z-values.
}
\examples{
p2z(p = c(0.005, 0.01, 0.05))
p2z(p = c(0.005, 0.01, 0.05), alternative = "greater")
p2z(p = c(0.005, 0.01, 0.05), alternative = "less")

p <- seq(0.001, 0.05, 0.0001)
plot(p, p2z(p), type = "l", ylim = c(0, 3.5), ylab = "z")
lines(p, p2z(p, alternative = "greater"), lty = 2)
legend("bottomleft", c("two-sided", "greater"), lty = c(1, 2), bty = "n")
}
\seealso{
\code{\link{ci2se}}, \code{\link{ci2estimate}}, \code{\link{ci2z}}, \code{\link{ci2p}}, \code{\link{z2p}}
}
