% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pValFunPlot.R
\name{pValFunPlot}
\alias{pValFunPlot}
\title{Plots the p-value function}
\usage{
pValFunPlot(
  thetao,
  thetar,
  seo,
  ser,
  c = NA,
  levelCI = 0.95,
  xlim = c(min(0, c(thetao, thetar) - 2 * c(seo, ser)), max(0, c(thetao, thetar) + 2 *
    c(seo, ser)))
)
}
\arguments{
\item{thetao}{Effect estimate of the original study}

\item{thetar}{Effect estimate of the replication study}

\item{seo}{Standard error of thetao}

\item{ser}{Standard error of thetar}

\item{c}{Default is variance ratios of the original and replication
effect estimates. This is usually the ratio of the sample
size of the replication study to the sample size of the
original study. It is also possible to choose c as a free parameter,
not related to the variance ratio.}

\item{levelCI}{Level of the confidence interval to be added on the plot}

\item{xlim}{limits of the x-axis of the plot}
}
\value{
\code{pValFunPlot} plots the p-value function
}
\description{
Plots the p-value function based on the effect estimates 
and standard errors of the original and the replication study.
}
\details{
the function also returns the minimum p-value, the 
Q-test p-value and the two-sided p-value
}
\examples{
pValFunPlot(thetao = 2, thetar = 1.9, seo = 1, ser = 0.5, levelCI = 0.95)
}
\author{
Leonhard Held, Florian Gerber, Charlotte Micheloud
}
