% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{RPCB}
\alias{RPCB}
\title{Data from the Reproducibility Project: Cancer Biology}
\format{
A \code{data.frame} with 112 rows and 42 variables
}
\source{
https://osf.io/squy7/
}
\usage{
data(RPCB)
}
\description{
Data from the \emph{Reproducibility Project Cancer Biology}
(RPCB). The variables are as follows:
\describe{
\item{\code{osf}}{Link to the OSF repository.}
\item{\code{id}}{???}
\item{\code{paper}}{???}
\item{\code{experiment}}{???}
\item{\code{effectType}}{Type of the effect.}
\item{\code{effect}}{???}
\item{\code{resulto}}{The direction of effect in the original study.}
\item{\code{no}}{The sample size in the original study.}
\item{\code{smdo}}{The standardized mean difference in the original study.}
\item{\code{so}}{The standard error in the original study.}
\item{\code{zo}}{The z-value in the original study.}
\item{\code{po}}{The two-sided p-value in the original study.}
\item{\code{po1}}{The one-sided p-value in the original study. It is only
available for studies where \code{resulto} is \code{"Positive"}.}
\item{\code{CIo_l}}{The lower bound of the 95\%-confidence interval for the
smd in the original study.}
\item{\code{CIo_u}}{The upper bound of the 95\%-confidence interval for the
smd in the original study.}
\item{\code{ESo}}{The effect size in the original study.
??? is this correct???}
\item{\code{seESo}}{The standard error of the effect size in the original
study.}
\item{\code{lowerESo}}{The lower bound of the 95\%-confidence interval for
the effect size in the original study.}
\item{\code{upperESo}}{The upper bound of the 95\%-confidence interval for
the effect size in the original study.}
\item{\code{resultr}}{The direction of effect in the replication study.}
\item{\code{nr}}{The sample size in the replication study.}
\item{\code{smdr}}{The standardized mean difference in the replication
study.}
\item{\code{sr}}{The standard error in the replication study.}
\item{\code{zr}}{The z-value in the replication study.}
\item{\code{pr}}{The two-sided p-value in the replication study.}
\item{\code{pr1}}{The one-sided p-value in the replication study. It
it is only available for studies where \code{resulto} is \code{"Positive"}.}
\item{\code{CIr_l}}{The lower bound of the 95\%-confidence interval for the
smd in the replication study.}
\item{\code{CIr_u}}{The lower bound of the 95\%-confidence interval for the
smd in the replication study.}
\item{\code{ESr}}{The effect size in the replication study.
???is this correct???}
\item{\code{seESr}}{The standard error of the effect size in the replication
study.}
\item{\code{lowerESr}}{The lower bound of the 95\%-confidence interval for
the effect size in the replication study.}
\item{\code{upperESr}}{The upper bound of the 95\%-confidence interval for
the effect size in the replication study.}
\item{\code{d}}{The relative effect size, i.e. \code{smdr/smdo}.}
\item{\code{c}}{The variance ratio, i.e. \code{so^2 / sr^2}.}
\item{\code{ttr}}{The significance criterion. ??? should we keep this ???.}
\item{\code{ma}}{The meta-analysis effect estimate. ???}
\item{\code{se_ma}}{The standard error for the meta-analysis effect
estimate. ???}
\item{\code{p_ma}}{The two-sided p-value for the meta-analysis effect
estimate. ???}
\item{\code{CIma_l}}{The lower bound of the 95\%-confidence interval for the
meta-analysis effect estimate. ???}
\item{\code{CIma_u}}{The upper bound of the 95\%-confidence interval for the
meta-analysis effect estimate. ???}
\item{\code{pS}}{The two-sided sceptical p-value for the meta-analysis. ???}
\item{\code{pS1}}{The one-sided sceptical p-value for the meta-analysis. ???}
}
}
\details{
This data set is a subset(???) of the data from the Reproducibility
Project: Cancer Biology (RPCB). TODO: Describe why we did not use their data
directly and why we recalculated all the stuff above.
}
\references{
Errington, T. M., Mathur, M., Soderberg, C. K., Denis, A.,
    Perfito, N., Iorns, E., & Nosek, B. A. (2021). Investigating the
    replicability of preclinical cancer biology. \emph{eLife}, \bold{10},
    e71601. https://doi.org/10.7554/eLife.71601 \cr\cr
Errington, T. M., Mathur, M., Soderberg, C. K., & Denis, A. (2021).
    Meta-analysis paper figures and tables.
    https://doi.org/10.17605/OSF.IO/SQUY7
}
\keyword{data}
