% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scepticalCI.R
\name{scepticalCI}
\alias{scepticalCI}
\title{Computes the sceptical confidence interval}
\usage{
scepticalCI(
  thetao,
  thetar,
  seo,
  ser,
  c = NA,
  levelCI = 0.95,
  alternative = c("two.sided", "greater", "less")
)
}
\arguments{
\item{thetao}{Effect estimate of the original study}

\item{thetar}{Effect estimate of the replication study}

\item{seo}{Standard error of thetao}

\item{ser}{Standard error of thetar}

\item{c}{Default is variance ratios of the original and replication
effect estimates. This is usually the ratio of the sample
size of the replication study to the sample size of the
original study. It is also possible to choose c as a free parameter,
not related to the variance ratio.}

\item{levelCI}{Level of the confidence interval.}

\item{alternative}{Either "two.sided" (default), "greater" or "less".
Specifies the type of the confidence interval and of the confidence level.}
}
\value{
\code{scepticalCI} returns the sceptical confidence interval and 
the minimum p-value of the p-value function between the original 
and replication effect estimates. The confidence interval splits 
into two if the minimum p-value is smaller than 1 - \code{levelCI}.
}
\description{
Computes the sceptical confidence interval based on the effect estimates 
and standard errors of the original and the replication study.
}
\details{
If the two-sided levelCI-confidence interval does not include 0, 
the two-sided controlled sceptical p-value will be smaller than sqrt(1 - levelCI).
If their is conflict between the original and the replication effect 
estimates, the confidence region splits into two disjoint intervals.
}
\examples{
scepticalCI(thetao = 2, thetar = 1.9, seo = 1, ser = 0.5, alternative = "two.sided", levelCI = 0.95)
}
\references{
Held, L., Micheloud, C., Balabdaoui, F. (2022).
 A Statistical Framework for Replicability. 
 \url{https://arxiv.org/abs/2207.00464}
}
\author{
Leonhard Held, Florian Gerber, Charlotte Micheloud
}
