% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Soto2019}
\alias{Soto2019}
\title{Data from the LOOPR Project}
\format{
A data frame with 143 rows and 15 variables
}
\source{
The relevant files were downloaded from \url{https://osf.io/d3xb7}
    on December 08, 2023 via the R script "Soto2019.R", which also executes
    the code to assemble the desired variables.
}
\usage{
data(Soto2019)
}
\description{
Data from the \emph{Life Outcomes of Personality Replication}
(LOOPR) project. The variables are as follows:
\describe{
\item{\code{outcome}}{The life outcome}
\item{\code{trait_predicted}}{The personality trait}
\item{\code{study}}{Study identifier, usually names of authors
from original study}
\item{\code{project}}{Name of replication project}
\item{\code{ro}}{Effect estimate of original study on correlation scale}
\item{\code{rr}}{Effect estimate of replication study on correlation scale}
\item{\code{fiso}}{Effect estimate of original study transformed to
Fisher-z scale}
\item{\code{fisr}}{Effect estimate of replication study transformed
to Fisher-z scale}
\item{\code{se_fiso}}{Standard error of Fisher-z transformed effect estimate
of original study}
\item{\code{se_fisr}}{Standard error of Fisher-z transformed effect estimate
of replication study}
\item{\code{po}}{Two-sided p-value from significance test of effect estimate
from original study}
\item{\code{pr}}{Two-sided p-value from significance test of effect estimate
from replication study}
 \item{\code{po1}}{One-sided p-value from significance test of effect
estimate from original study (in the direction of the original effect
estimate)}
\item{\code{pr1}}{One-sided p-value from significance test of effect estimate
from replication study (in the direction of the original effect estimate)}
\item{\code{pm_belief}}{For this data set, this variable is always \code{NA}.
It is only there for completeness, such that the RProjects and Soto2019 data
sets can easily be bound together.}
\item{\code{no}}{Sample size in original study. Sometimes the values in this
variable are not integers. This happens if the trait-outcome associations are
composed from multiple original studies.}
\item{\code{nr}}{Sample size in replication study. Sometimes the values in
this variable are not integers. This happens if the trait-outcome
associations are composed from multiple original studies.}
}
}
\details{
This data set corresponds largely to the data presented in Table 1
in Soto (2019). However, we have attached the names of the authors of the
original study from the data available through the OSF repository.
}
\references{
Soto, C. J. (2019). How Replicable Are Links Between Personality
    Traits and Consequential Life Outcomes? The Life Outcomes of Personality
    Replication Project. \emph{Psychological Science}, \bold{30}(5), 711-727.
    https://doi.org/10.1177/0956797619831612 \cr\cr
Soto, C. J. (2019, May 28). The Life Outcomes Of Personality Replication
    Project. Retrieved from osf.io/d3xb7
}
\keyword{data}
