% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversionHelpers.R
\name{z2p}
\alias{z2p}
\title{Transforms z-values into p-values.}
\usage{
z2p(z, alternative = "two.sided")
}
\arguments{
\item{z}{Numeric vector of z-values.}

\item{alternative}{Direction of the alternative of the p-value.
Either "one.sided" (default), "two.sided", "less", or "greater".}
}
\value{
A Numeric vector of p-values.
}
\description{
Transforms z-values into p-values.
}
\examples{
z2p(z = c(1, 2, 5))
z2p(z = c(1, 2, 5), alternative = "less")
z2p(z = c(1, 2, 5), alternative = "greater")

z <- seq(-3, 3, by = 0.01)
plot(z, z2p(z), type = "l", xlab = "z", ylab = "p", ylim = c(0, 1))
lines(z, z2p(z, alternative = "greater"), lty = 2)
legend("topright", c("two-sided", "greater"), lty = c(1, 2), bty = "n")
}
\seealso{
\code{\link{ci2se}}, \code{\link{ci2estimate}}, \code{\link{ci2z}}, \code{\link{ci2p}}, \code{\link{p2z}}
}
